requirements.txt
Django
djangorestframework
django-filter
djangorestframework-simplejwt
Pillow